% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C02_fit_CARbym.R
\name{fit_CARbym}
\alias{fit_CARbym}
\title{Fit a Bivariate spatial generalized linear mixed model with BYM structured random effects}
\usage{
fit_CARbym(
  formula,
  data = NULL,
  trials,
  W,
  burnin = 5000,
  n_sample = 10000,
  thin = 5,
  prior_beta_mean = NULL,
  prior_beta_var = NULL,
  prior_Tau_df = NULL,
  prior_Tau_scale = NULL,
  prior_Sigma_df = NULL,
  prior_Sigma_scale = NULL,
  verbose = TRUE
)
}
\arguments{
\item{formula}{A formula for the covariate part of the model using the syntax of the lm() function. Offsets can be included here using the offset() function.  The covariates should each be a K*1 vector, where K is the number of spatial units. The response variable can contain missing values (NA).}

\item{data}{An optional data frame containing the variables in the model. If not found in data, the variables are taken from the environment from which \code{fit_CARbym} is called.}

\item{trials}{A matrix representing the number of trials for the binomial model, with dimensions matching those of the response variable.}

\item{W}{A spatial weight matrix representing neighborhood structures. which should be a non-negative K by K neighbourhood matrix (where K is the number of spatial units). Typically a binary specification is used, where the (j,k)th element equals one if areas (j, k) are spatially close (e.g. share a common border) and is zero otherwise. The matrix can be non-binary, but each row must contain at least one non-zero entry.}

\item{burnin}{The number of burn-in iterations for the MCMC sampling (default is 5000).}

\item{n_sample}{The total number of MCMC samples (default is 10000).}

\item{thin}{The thinning interval for the MCMC samples (default is 5).}

\item{prior_beta_mean}{A vector of prior means for the regression parameters beta (Gaussian priors are assumed). Defaults to a vector of zeros.}

\item{prior_beta_var}{A vector of prior variances for the regression parameters beta (Gaussian priors are assumed). Defaults to a vector with values 100000.}

\item{prior_Tau_df}{The prior degrees of freedom for the Inverse-Wishart prior for Tau. Defaults to J+1.}

\item{prior_Tau_scale}{The prior J times J scale matrix for the Inverse-Wishart prior for Tau. Defaults to the identity matrix divided by 1000.}

\item{prior_Sigma_df}{The prior degrees of freedom for the Inverse-Wishart prior for Sigma. Defaults to J+1.}

\item{prior_Sigma_scale}{The prior J times J scale matrix for the Inverse-Wishart prior for Sigma. Defaults to the identity matrix divided by 1000.}

\item{verbose}{Logical; if TRUE, prints progress during the MCMC sampling (default is TRUE).}
}
\value{
A list containing the model fit, including posterior samples, fitted values, residuals, and model summary.
\describe{
\item{summary.results }{A summary table of the parameters.}
\item{samples }{A list containing the MCMC samples, as well as the loglik of each samples from the model.}
\item{fitted.values }{A matrix of fitted values for each area and responsevariable.}
\item{residuals }{A list with 2 elements, where each element is a vector of a type of residuals. The types of residual are "response" (raw), and "pearson".}
\item{modelfit }{Model fit criteria including the Deviance Information Criterion(DIC) and its corresponding estimated effective number of parameters (p.d), the Log Marginal Predictive Likelihood (LMPL), the Watanabe-Akaike Information Criterion (WAIC) and its corresponding estimated number of effective parameters (p.w), the Leave-One-Out Information Criterion (LOOIC), the Expected Log Predictive Density (ELPD), and the loglikelihood.}
\item{accept }{The acceptance probabilities for the parameters.}
\item{localised.structure }{NULL, for compatability with other models.}
\item{formula }{The formula (as a text string) for the response, covariate and offset parts of the model}
\item{model }{A text string describing the model fit.}
\item{X }{The design matrix of covariates.}
}
}
\description{
This function fits a bivariate generalized linear mixed model to areal unit data, where the response variable can be binomial. The linear predictor is modelled by known covariates and 2 vectors of random effects. The latter are modelled by the BYM conditional autoregressive prior proposed by Besag et al. (1991), and further details are given in the vignette accompanying this package. Inference is conducted in a Bayesian setting using Markov chain Monte Carlo (MCMC) simulation. Missing (NA) values are allowed in the response, and posterior predictive distributions are created for the missing values using data augmentation. These are saved in the "samples" argument in the output of the function and are denoted by "Y".
}
\references{
\describe{
Besag, J., J. York, and A. Mollie (1991). Bayesian image restoration with two applications in spatial statistics. Annals of the Institute of Statistics and Mathematics 43, 1-59.
}
}
