% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D02_fit_CARbym_copula.R
\name{fit_CARbym_copula}
\alias{fit_CARbym_copula}
\title{Fit a Bivariate Copula CAR Bayesian Model}
\usage{
fit_CARbym_copula(
  formula,
  copula = "gaussian",
  data = NULL,
  trials,
  W,
  burnin = 5000,
  n_sample = 10000,
  thin = 5,
  prior_beta_mean = NULL,
  prior_beta_var = NULL,
  prior_tau2 = NULL,
  prior_sigma2 = NULL,
  tau2_AB = NULL,
  sigma2_AB = NULL,
  verbose = TRUE
)
}
\arguments{
\item{formula}{A symbolic description of the model to be fitted.}

\item{copula}{The type of copula to use (default is "gaussian"). Currently, only "gaussian" is supported.}

\item{data}{An optional data frame containing the variables in the model.}

\item{trials}{The number of trials for the Binomial model.}

\item{W}{A spatial weight matrix.}

\item{burnin}{The number of burn-in iterations for the MCMC sampling (default is 5000).}

\item{n_sample}{The total number of MCMC samples (default is 10000).}

\item{thin}{The thinning interval for the MCMC samples (default is 5).}

\item{prior_beta_mean}{Prior mean for the regression coefficients (optional).}

\item{prior_beta_var}{Prior variance for the regression coefficients (optional).}

\item{prior_tau2}{Prior for the variance parameter tau^2 (optional).}

\item{prior_sigma2}{Prior for the variance parameter sigma^2 (optional).}

\item{tau2_AB}{Hyperparameters for the prior distribution of tau^2 (optional).}

\item{sigma2_AB}{Hyperparameters for the prior distribution of sigma^2 (optional).}

\item{verbose}{Logical; if TRUE, prints progress during the MCMC sampling (default is TRUE).}
}
\value{
A model object containing the results of the fitted CAR BYM model with the specified copula.
}
\description{
This function fits a bivariate copula CAR Bayesian model with a BYM structure
for a Binomial distribution.
Currently, it supports only the Gaussian copula.
}
