% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D01_fit_CARleroux_copula.R
\name{fit_CARleroux_copula}
\alias{fit_CARleroux_copula}
\title{Fit a Bivariate Copula CAR Bayesian Model}
\usage{
fit_CARleroux_copula(
  formula,
  copula = "gaussian",
  data = NULL,
  trials,
  W,
  burnin = 5000,
  n_sample = 10000,
  thin = 5,
  prior_beta_mean = NULL,
  prior_beta_var = NULL,
  prior_tau2 = NULL,
  alpha = NULL,
  rho = NULL,
  verbose = TRUE
)
}
\arguments{
\item{formula}{The model formula.}

\item{copula}{The type of copula to use (e.g., "gaussian", "clayton").}

\item{data}{The dataset to be used.}

\item{trials}{The number of trials for the Binomial model.}

\item{W}{Spatial weight matrix.}

\item{burnin}{Number of burn-in iterations for the MCMC.}

\item{n_sample}{Total number of MCMC samples.}

\item{thin}{Thinning interval for the MCMC.}

\item{prior_beta_mean}{Prior mean for the regression coefficients.}

\item{prior_beta_var}{Prior variance for the regression coefficients.}

\item{prior_tau2}{Prior for the variance parameter.}

\item{alpha}{Copula parameter alpha.}

\item{rho}{Spatial autocorrelation parameter.}

\item{verbose}{Logical; if TRUE, prints progress during the MCMC sampling.}
}
\value{
A model object containing the results of the fitted model.
}
\description{
This function fits a bivariate copula CAR Bayesian model with a Leroux structure
for a Binomial distribution.
}
