% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F03_fitted.CARBayes.R
\name{fitted.CARBayes}
\alias{fitted.CARBayes}
\title{Fitted Values for CARBayes Models}
\usage{
\method{fitted}{CARBayes}(object, ...)
}
\arguments{
\item{object}{An object of class \code{CARBayes}.}

\item{...}{Additional arguments passed to or from other methods.}
}
\value{
A matrix or list of fitted values, depending on the model type (univariate or multivariate).
}
\description{
This function returns the fitted values for objects of class \code{CARBayes}.
}
