% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F02_residuals.CARBayes.R
\name{residuals.CARBayes}
\alias{residuals.CARBayes}
\title{Residuals for CARBayes Models}
\usage{
\method{residuals}{CARBayes}(object, type = "pearson", ...)
}
\arguments{
\item{object}{An object of class \code{CARBayes}.}

\item{type}{The type of residuals to return. Options are \code{"pearson"} (default) and \code{"response"}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
The residuals of the specified type. For multivariate models, the function returns a list of residuals; for univariate models, it returns a matrix of residuals.
}
\description{
This function computes and returns residuals for objects of class \code{CARBayes}.
}
