% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C02_fit_CARbym.R
\name{fit_CARbym}
\alias{fit_CARbym}
\title{Fit a Bivariate BYM CAR Model for Binomial Data}
\usage{
fit_CARbym(
  formula,
  data = NULL,
  trials,
  W,
  burnin = 5000,
  n_sample = 10000,
  thin = 5,
  prior_beta_mean = NULL,
  prior_beta_var = NULL,
  prior_Tau_df = NULL,
  prior_Tau_scale = NULL,
  prior_Sigma_df = NULL,
  prior_Sigma_scale = NULL,
  verbose = TRUE
)
}
\arguments{
\item{formula}{A symbolic description of the model to be fitted.}

\item{data}{An optional data frame containing the variables in the model. If not found in data, the variables are taken from the environment from which \code{BVS_CARbym_Binomial} is called.}

\item{trials}{A matrix representing the number of trials for the binomial model, with dimensions matching those of the response variable.}

\item{W}{A spatial weight matrix representing neighborhood structures.}

\item{burnin}{The number of burn-in iterations for the MCMC sampling (default is 5000).}

\item{n_sample}{The total number of MCMC samples (default is 10000).}

\item{thin}{The thinning interval for the MCMC samples (default is 5).}

\item{prior_beta_mean}{A vector of prior means for the regression coefficients. Defaults to a vector of zeros.}

\item{prior_beta_var}{A vector of prior variances for the regression coefficients. Defaults to a large value (100000).}

\item{prior_Tau_df}{Degrees of freedom for the prior distribution of the variance-covariance matrix Tau. Defaults to \code{J + 1}, where \code{J} is the number of columns in the response variable.}

\item{prior_Tau_scale}{A scale matrix for the prior distribution of Tau. Defaults to a small diagonal matrix.}

\item{prior_Sigma_df}{Degrees of freedom for the prior distribution of the variance-covariance matrix Sigma. Defaults to \code{J + 1}.}

\item{prior_Sigma_scale}{A scale matrix for the prior distribution of Sigma. Defaults to a small diagonal matrix.}

\item{verbose}{Logical; if TRUE, prints progress during the MCMC sampling (default is TRUE).}
}
\value{
A list containing the model fit, including posterior samples, fitted values, residuals, and model summary.
}
\description{
This function fits a bivariate Conditional Autoregressive (CAR) BYM model for binomial data. It allows for spatial correlation in the response using the Besag-York-Mollié (BYM) model.
}
